Split test per type of connection, so it's more transparent when new APIs are added. Update README.md (refresh link to the latest Anaplan Connect documentation and change download file example after latest update). Finalize release v0.5.0.
