+++ apapi/__version__.py
@@ -1,7 +1,7 @@
 __title__ = "apapi"
 __description__ = "Unofficial Anaplan API Python client library"
 __url__ = "https://github.com/DLZaan/apapi"
-__version__ = "0.4.0"  # also needs to be bumped in docs/module.html.jinja2
+__version__ = "0.5.0"  # also needs to be bumped in docs/module.html.jinja2
 __author__ = "Milosz Kierepka"
 __author_email__ = "milosz.kierepka+work@gmail.com"
 __license__ = "Apache 2.0"
