+++ docs/module.html.jinja2
@@ -1,4 +1,4 @@
 {% extends "default/module.html.jinja2" %}
 {% block nav_footer %}
-    <footer>Latest Release: v0.4.0</footer>
+    <footer>Latest Release: v0.5.0</footer>
 {% endblock %}
